# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "An API for users to register/login and view different products.\n                    Registered users can add products to their cart and checkout to create \n                    their order. Registered users with admin role can add, edit and delete the products."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "Get all users orders",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched list of orders",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Orders found"
                    },
                    "orders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/orders/admin": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "Get all orders. Admins only",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched list of orders",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Orders found"
                    },
                    "orders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "403": {
            "description": "User is not admin",
            "schema": {
              "$ref": "#components/schemas/NotAdmin"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/orders/{id}": {
      "put": {
        "tags": [
          "Orders"
        ],
        "description": "Update status order by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the order to update",
            "example": 1
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Status of order to update",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "New status"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Status updated"
                    },
                    "product": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request — missing required fields.",
            "schema": {
              "$ref": "#components/schemas/MissingParameters"
            }
          },
          "401": {
            "description": "User provided invalid token",
            "schema": {
              "$ref": "#components/schemas/FailedAuthentication"
            }
          },
          "403": {
            "description": "User is not admin",
            "schema": {
              "$ref": "#components/schemas/NotAdmin"
            }
          },
          "404": {
            "description": "Order not found",
            "schema": {
              "$ref": "#components/schemas/ResourceNotFound"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#components/schemas/ServerError"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "ordernumber": {
            "type": "string",
            "example": "djeivwhf4"
          },
          "status": {
            "type": "string",
            "example": "In progress"
          },
          "discount": {
            "type": "integer",
            "example": 0
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "membership_id": {
            "type": "integer",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-02T09:25:53Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-02T09:25:53Z"
          }
        }
      },
...
    }
  }
}